# Conflicts:
#	CPU/CPU.c
#	main.c

+++ ALU/ALU.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include "ALU.h"
-#include "ALUHelpers/ALUhelpers.h"
+#include "ALUHelpers/ALUhelpers.c"
 
-
-int ALU(uint8_t opcode, int32_t operand_A, int32_t operand_B, int *flag_out) {
+// git test
+int ALU(int opcode, int32_t operand_A, int32_t operand_B, int *flag_out) {
     int out;
 
     if (opcode == ALU_NOP)
